# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MailCue API",
    "description": "Realistic email testing server REST API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/emails": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "List All Emails",
        "description": "List emails in a mailbox with pagination, search, and sorting.\n\nEmails are fetched directly from IMAP. The ``search`` parameter\nmaps to IMAP ``TEXT`` search which covers subject and body.",
        "operationId": "list_all_emails_api_v1_emails_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "mailbox",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target mailbox address (user@domain)",
              "title": "Mailbox"
            },
            "description": "Target mailbox address (user@domain)"
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "IMAP folder name",
              "default": "INBOX",
              "title": "Folder"
            },
            "description": "IMAP folder name"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text search query",
              "title": "Search"
            },
            "description": "Full-text search query"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort order (date_asc, date_desc)",
              "default": "date_desc",
              "title": "Sort"
            },
            "description": "Sort order (date_asc, date_desc)"
          },
          {
            "name": "thread_view",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Group conversations: sort the page by (thread_id, date asc).",
              "default": false,
              "title": "Thread View"
            },
            "description": "Group conversations: sort the page by (thread_id, date asc)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emails/{uid}": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get Single Email",
        "description": "Fetch a single email by its IMAP UID with full body and headers.",
        "operationId": "get_single_email_api_v1_emails__uid__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          },
          {
            "name": "mailbox",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target mailbox address",
              "title": "Mailbox"
            },
            "description": "Target mailbox address"
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "INBOX",
              "title": "Folder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Emails"
        ],
        "summary": "Delete Single Email",
        "description": "Delete an email by UID (sets \\Deleted flag and expunges).",
        "operationId": "delete_single_email_api_v1_emails__uid__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          },
          {
            "name": "mailbox",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target mailbox address",
              "title": "Mailbox"
            },
            "description": "Target mailbox address"
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "INBOX",
              "title": "Folder"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emails/{uid}/raw": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Get Raw Email",
        "description": "Download the raw RFC 5322 source of an email as a ``.eml`` file.",
        "operationId": "get_raw_email_api_v1_emails__uid__raw_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          },
          {
            "name": "mailbox",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target mailbox address",
              "title": "Mailbox"
            },
            "description": "Target mailbox address"
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "INBOX",
              "title": "Folder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emails/{uid}/attachments/{part_id}": {
      "get": {
        "tags": [
          "Emails"
        ],
        "summary": "Download Attachment",
        "description": "Download a specific MIME attachment identified by its part ID.",
        "operationId": "download_attachment_api_v1_emails__uid__attachments__part_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          },
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Part Id"
            }
          },
          {
            "name": "mailbox",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target mailbox address",
              "title": "Mailbox"
            },
            "description": "Target mailbox address"
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "INBOX",
              "title": "Folder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emails/send": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Send New Email",
        "description": "Send an email via the local SMTP server (Postfix).\n\nThe sender address (``from_address``) must belong to the\nauthenticated user's mailbox (admins may send from any).",
        "operationId": "send_new_email_api_v1_emails_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Send New Email Api V1 Emails Send Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/emails/inject": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Inject Single Email",
        "description": "Inject an email directly into a mailbox via IMAP APPEND.\n\n**Admin only.** Bypasses SMTP delivery entirely -- the email\nappears in the target mailbox immediately. Ideal for test data setup.",
        "operationId": "inject_single_email_api_v1_emails_inject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InjectEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Inject Single Email Api V1 Emails Inject Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/emails/bulk-inject": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Bulk Inject Emails",
        "description": "Inject multiple emails into mailboxes in a single request.\n\n**Admin only.** Each email in the ``emails`` array is injected\nindependently. Partial failures are reported in the response\nwithout aborting the entire batch.",
        "operationId": "bulk_inject_emails_api_v1_emails_bulk_inject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkInjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkInjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/emails/validate": {
      "post": {
        "tags": [
          "Emails"
        ],
        "summary": "Validate Email Endpoint",
        "description": "Validate email address format, DNS domain MX/NS, SMTP availability, and disposable status.",
        "operationId": "validate_email_endpoint_api_v1_emails_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/api/v1/mailboxes/{mailbox_address}/emails": {
      "get": {
        "tags": [
          "Mailboxes"
        ],
        "summary": "List Mailbox Emails",
        "description": "List emails for a specific mailbox (path-based variant).",
        "operationId": "list_mailbox_emails_api_v1_mailboxes__mailbox_address__emails_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "mailbox_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mailbox Address"
            }
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "INBOX",
              "title": "Folder"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "date_desc",
              "title": "Sort"
            }
          },
          {
            "name": "thread_view",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "When true, sort the page by (thread_id, date asc) so conversations render as contiguous groups.",
              "default": false,
              "title": "Thread View"
            },
            "description": "When true, sort the page by (thread_id, date asc) so conversations render as contiguous groups."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mailboxes/{mailbox_address}/emails/{uid}": {
      "get": {
        "tags": [
          "Mailboxes"
        ],
        "summary": "Get Mailbox Email",
        "description": "Fetch a single email by UID from a specific mailbox.",
        "operationId": "get_mailbox_email_api_v1_mailboxes__mailbox_address__emails__uid__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "mailbox_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mailbox Address"
            }
          },
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "INBOX",
              "title": "Folder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Mailboxes"
        ],
        "summary": "Delete Mailbox Email",
        "description": "Delete an email by UID from a specific mailbox.",
        "operationId": "delete_mailbox_email_api_v1_mailboxes__mailbox_address__emails__uid__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "mailbox_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mailbox Address"
            }
          },
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "INBOX",
              "title": "Folder"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkInjectRequest": {
        "properties": {
          "emails": {
            "items": {
              "$ref": "#/components/schemas/InjectEmailRequest"
            },
            "type": "array",
            "title": "Emails"
          }
        },
        "type": "object",
        "required": [
          "emails"
        ],
        "title": "BulkInjectRequest",
        "description": "Inject multiple emails at once."
      },
      "BulkInjectResponse": {
        "properties": {
          "injected": {
            "type": "integer",
            "title": "Injected"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "injected",
          "failed",
          "ids"
        ],
        "title": "BulkInjectResponse",
        "description": "Result of a bulk inject operation."
      },
...
      "EmailListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "emails": {
            "items": {
              "$ref": "#/components/schemas/EmailSummary"
            },
            "type": "array",
            "title": "Emails"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "emails"
        ],
        "title": "EmailListResponse",
        "description": "Paginated list of email summaries."
      },
...
      "EmailValidationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 1,
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "EmailValidationRequest",
        "description": "Request body for email address validation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}